Organize the structure of the files
