+++ app/components/listings/ListingReservation.tsx
@@ -2,7 +2,7 @@
 
 import { Range } from "react-date-range";
 
-import Button from "../Button";
+import Button from "../elementsUi/Button";
 import Calendar from "../inputs/Calendar";
 
 interface ListingReservationProps {
