+++ app/listings/[listingId]/page.tsx
@@ -3,8 +3,8 @@ import getCurrentUser from "@/app/actions/getCurrentUser";
 import getListingById from "@/app/actions/getListingById";
 import getReservations from "@/app/actions/getReservations";
 
-import ClientOnly from "@/app/components/ClientOnly";
-import EmptyState from "@/app/components/EmptyState";
+import ClientOnly from "@/app/components/elementsUi/ClientOnly";
+import EmptyState from "@/app/components/elementsUi/EmptyState";
 
 import ListingClient from "./ListingClient";
