+++ app/page.tsx
@@ -1,12 +1,12 @@
-import Container from "@/app/components/Container";
+import Container from "@/app/components/elementsUi/Container";
 import ListingCard from "@/app/components/listings/ListingCard";
-import EmptyState from "@/app/components/EmptyState";
+import EmptyState from "@/app/components/elementsUi/EmptyState";
 
 import getListings, { 
   IListingsParams
 } from "@/app/actions/getListings";
 import getCurrentUser from "@/app/actions/getCurrentUser";
-import ClientOnly from "./components/ClientOnly";
+import ClientOnly from "./components/elementsUi/ClientOnly";
 
 interface HomeProps {
   searchParams: IListingsParams
