+++ app/properties/page.tsx
@@ -1,6 +1,6 @@
 
-import EmptyState from "@/app/components/EmptyState";
-import ClientOnly from "@/app/components/ClientOnly";
+import EmptyState from "@/app/components/elementsUi/EmptyState";
+import ClientOnly from "@/app/components/elementsUi/ClientOnly";
 
 import getCurrentUser from "@/app/actions/getCurrentUser";
 import getListings from "@/app/actions/getListings";
