+++ app/reservations/ReservationsClient.tsx
@@ -7,8 +7,8 @@ import { useRouter } from "next/navigation";
 
 import { SafeReservation, SafeUser } from "@/app/types"
 ;
-import Heading from "@/app/components/Heading";
-import Container from "@/app/components/Container";
+import Heading from "@/app/components/elementsUi/Heading";
+import Container from "@/app/components/elementsUi/Container";
 import ListingCard from "@/app/components/listings/ListingCard";
 
 interface ReservationsClientProps {
